Refactor code to streamline initialization of view models
The changes mainly involve refactoring the MultipleChoiceSettingViewModel and TextSettingViewModel in order to simplify their initialization process. A constructor parameter `initialValue` is added to the MultipleChoiceSettingViewModel to directly set the initial value rather than through a property. Also, the way that settings view model is initialized in SettingsModel has been simplified. Lastly, in TextSettingViewModel, a .Skip(1) is added to the value listener to prevent invoking the setting change handling prematurely.