Refactor PulsarWorker.Desktop to optimise loading of clusters
This commit involves a large-scale refactor of the PulsarWorker.Desktop, focusing on the ViewModel classes and related files regarding the loading of clusters. The primary objective of this update is to optimise the application's performance by streamlining and simplifying the said operation.

A couple of ViewModel files were deleted, and we added new ViewModel files for Cluster, NameSpace, Topic, and Tenant to better handle the loading of clusters. PulsarModel was created to replace the PulsarTreeModel and ClusterViewModel, a component of the new architecture, is now handling cluster loading and displaying.

Moreover, significant changes were made in PulsarApiViewModel, which now leans on the newly established ClusterViewModel to load and display clusters.

Finally, reactive activation and deactivation were included in ViewModelBase, with updates being made to the corresponding ViewModels to account for these changes.

Overall, this update simplifies the codebase, makes it easier to maintain, and improves the application's efficiency.