Update Tenant and Cluster view, Add tenant retrieval
Updated the TenantView and ClusterView to display the new data structure. A StackPanel has been added to show NameSpaces(Tenants) in the respective ViewModel. Implemented GetTenants function in PulsarModel to asynchronously retrieve the list of tenants. This allows the application to display an interactive and detailed view of tenants under each cluster. Also, updated the TenantViewModel and ClusterViewModel to support the new structure and data retrieval process.